water_py/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.4
django-cors-headers==3.2.1
django-import-export==2.0.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.1.0
et-xmlfile==1.0.1
jdcal==1.4.1
jiphy==1.2.2
MarkupPy==1.14
odfpy==1.4.1
Pillow==7.1.1
post==2019.4.13
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
pywin32==227
PyYAML==5.3.1
requests==2.23.0
sqlparse==0.3.1
tablib==1.1.0
urllib3==1.25.9
WMI==1.4.9
xlrd==1.2.0
xlwt==1.3.0
